+++ CHANGELOG.md
@@ -12,11 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk: add ability to do placeAndTake order with bracket orders attached
 - sdk: add option to cancel existing orders in market for place and take order
 - sdk: add option to get signed settlePnl tx back from a market order
-
 - program: auto derisk lp positions in settle pnl ([#766](https://github.com/drift-labs/protocol-v2/pull/766))
 - program: increase full perp liquidation threshold ([#807](https://github.com/drift-labs/protocol-v2/pull/807))
 - program: remove spot fee pool transfer ([#800](https://github.com/drift-labs/protocol-v2/pull/800))
 - program: increase insurance tier max ([#784](https://github.com/drift-labs/protocol-v2/pull/784))
+- sdk: can specify max custom margin ratio to initialize a new account with
 
 ### Fixes